Remove non-leaf conversion logic
If a leaf or non-leaf's children are all removed, the node
should be deleted. The one exception, and apparently the
reason for `_convertToLeafNode`, is if the entire tree is
emptied and we want the root to convert from non-leaf to
leaf. This is handled by `RTree._resetRoot`, therefore
we can remove `_convertToLeafNode`. This will additionally
resolve a bug where empty branches of the tree remained
because the conversion from non-leaf to leaf escaped the
`childrenToRemove` logic in `NonLeafNode.remove`.